requirements.txt
crewai[tools,anthropic]
fastapi
uvicorn
python-multipart
python-pptx
google-generativeai
gitpython
pydantic
python-dotenv
